[FIND-510, ISO-121]  Read JSON array usernames from HOD user metadata
Now that the Java Hod Client returns metadata as a JsonNode, I have updated here to extract the display name in the Hod Authentication Provider and store in a new object HodUserMetadata using the specific resolver.

The resolver interface now expects a map of String and JsonNode as the argument for the resolve() method.

[rev: matt.gordon]